Check master branch status before commit to gh-pages.

+++ gh-pages.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ -z $(git status | grep 'nothing to commit')  ]]; then
+    echo "Commit first! Abort."
+    exit 0
+fi
+
 export LANG=en; make clean; make html
 rm -rf /tmp/output/
 mv output/ /tmp/
